+++ client/src/store/slices/chatsSlice/chatsSlice.js
@@ -1,6 +1,5 @@
-import {createAsyncThunk, createSlice} from '@reduxjs/toolkit';
-import {convertRawStartChatsData} from "../../utils/convertRawStartChatsData";
-import {config} from "../../config";
+import {createSlice} from '@reduxjs/toolkit';
+import {loadOldMessages, loadStartChatMessages, setChatsData} from "./chatsThunks";
 
 
 class ChatsState {
@@ -20,26 +19,6 @@ class ChatsState {
     this.isOldMessagesLoading = isOldMessagesLoading;
   }
 }
-
-export const setChatsData = createAsyncThunk('chats/createChatsData', async (userId) => {
-  const {rawChatsInfo, lastMessages} = await (await fetch(`/user/${userId}/chatsData`)).json();
-  return convertRawStartChatsData(rawChatsInfo, lastMessages);
-});
-
-export const loadOldMessages = createAsyncThunk('chats/loadOldMessages', async ({chatId, offset, cbOnAllLoaded}) => {
-  const {LOAD_MESSAGES_THRESHOLD} = config;
-  const {oldMessages} = await (await fetch(`/api/chat/${chatId}/${offset}/${LOAD_MESSAGES_THRESHOLD}`)).json();
-  if (oldMessages.length < LOAD_MESSAGES_THRESHOLD) {
-    cbOnAllLoaded();
-  }
-
-  return {chatId, messages: oldMessages};
-});
-
-export const loadStartChatMessages = createAsyncThunk('chats/loadStartChatMessages',
-  async ({chatId, cbOnAllLoaded}) => {
-    return loadOldMessages({chatId, offset: 1, cbOnAllLoaded});
-});
 
 /**
  * @type ChatsState
