+++ client/src/store/slices/userSlice/userThunks.js
@@ -1,0 +1,6 @@
+import {createAsyncThunk} from "@reduxjs/toolkit";
+
+export const setCurrentUser = createAsyncThunk('user/setCurrentUser', async () => {
+  const {user} = await (await fetch('/user/self')).json();
+  return user;
+});
